Revert "Merge branch 'dev' into master"
This reverts commit f903a36d9083c880e6fb30a266553ee009e72bdd, reversing
changes made to 393db275042c2d8f4b98487aba09a67b4434b11b.

+++ modules/control/proc/marigold/__init__.py
@@ -1,4 +1,3 @@
-import torch
 from PIL import Image
 from modules.control.util import HWC3, resize_image
 from modules import devices
@@ -29,7 +28,7 @@ class MarigoldDetector:
         color_map: str = "Spectral",
         output_type=None,
     ):
-        self.model.to(device=devices.device, dtype=torch.float16)
+        self.model.to(device=devices.device, dtype=devices.dtype)
         res = self.model(
             input_image,
             denoising_steps=denoising_steps,
